Fixes bugs on reduce_distance_kernel and predict_query_values.
- reduce_distance_kernel had thread blocks overwriting elements
in the global_nn_dist and global_nn_idx arrays, thus they were not
able to run for the other (k-1) neighbors. On top of that, other
thread blocks (due to the nature of the reduction) where writing
to thise global arrays, because the new step was k instead of len.
The solution was to have one buffer where each thread reads the
data from and a separate buffer to write the reduced k neighbors.
The pointers for the two arrays will be swapped before proceding
to the next iteration.

- predict_query_values had a few indexing bugs, which are now
fixed.

NEXT STEPS:
- predict_query_values should distribute the work to more thread
blocks, maybe even move it outside of query blocking for loop.
- reduce_distance_kernel should work with the optimization where
each thread block is assigned 2 * blockDim.x distances, so all
of each threads run for one iteration.
- Try to get rid of the mod operation and the if threadIdx.x
statement in reduce_distance_kernel and thread_block_reduction.
- Introduce blocking for the training elements (2D grid(?))
- Write the debug file